refactor: 양수인 소수로도 덧셈이 가능하도록 리팩토링

+++ src/main/java/calculator/domain/Calculator.java
@@ -1,5 +1,5 @@
 package calculator.domain;
 
 public interface Calculator {
-    int calculate();
+    double calculate();
 }
